Without Stochasticity
The old version had a purely stochastic prediction of N forecast values and then a mean/std of the ensemble. But I realised that the equation can be clearly separated into a component for mean and standard deviation and hence, both can be computed separately without the need to construct ensembles of N points.